Limit average tram stop duration calculation to 1 day of data
Previously it was calculating the tram stop duration (i.e. the
tram passenger load duration at stops) based on the whole dataset,
which is very large.

Instead, we can calculate the average based on the recent 24h of data.
It's not as accurate, but it gets into the correct ballpark.